wdm-package {wdm}	R Documentation

Weighted Dependence Measures

Description

Provides efficient implementations of weighted dependence measures and related asymptotic tests for independence. Implemented measures are the Pearson correlation, Spearman's rho, Kendall's tau, Blomqvist's beta, and Hoeffding's D; see, e.g., Nelsen (2006) <doi:10.1007/0-387-28678-0> and Hollander et al. (2015, ISBN:9780470387375).

Details

The DESCRIPTION file: This package was not yet installed at build time.

Author(s)

Maintainer: Thomas Nagler mail@tnagler.com

See Also

Useful links:

https://github.com/tnagler/wdm-r

Report bugs at https://github.com/tnagler/wdm-r/issues

[Package wdm version 0.2.6 Index]