Independence Tests for Weighted Dependence Measures

Computes a (possibly weighted) dependence measure between x and y if these are vectors. If x and y are matrices then the measure between the columns of x and the columns of y are computed.

indep_test(
  x,
  y,
  method = "pearson",
  weights = NULL,
  remove_missing = TRUE,
  alternative = "two-sided"
)

Arguments

x, y
numeric vectors of data values. x and y must have the same length.

method
the dependence measure; see Details for possible values.

weights
an optional vector of weights for the observations.

remove_missing
if TRUE, all (pairswise) incomplete observations are removed; if FALSE, the function throws an error if there are incomplete observations.

alternative
indicates the alternative hypothesis and must be one of "two-sided", "greater" or "less". You can specify just the initial letter. "greater" corresponds to positive association, "less" to negative association.

Details

Available methods:

"pearson": Pearson correlation

"spearman": Spearman's \(\rho\)

"kendall": Kendall's \(\tau\)

"blomqvist": Blomqvist's \(\beta\)

"hoeffding": Hoeffding's \(D\)

Partial matching of method names is enabled. All methods except "hoeffding" work with discrete variables.

Examples

x <- rnorm(100)
y <- rpois(100, 1)  # all but Hoeffding's D can handle ties
w <- runif(100)

indep_test(x, y, method = "kendall")               # unweighted

#>      estimate  statistic   p_value n_eff  method alternative
#> 1 -0.01762804 -0.2129434 0.8313711   100 kendall   two-sided

indep_test(x, y, method = "kendall", weights = w)  # weighted

#>     estimate statistic   p_value    n_eff  method alternative
#> 1 0.01242062 0.1281651 0.8980183 74.08639 kendall   two-sided